@@ -27,12 +27,12 @@
     "axios": "^1.4.0",
     "clsx": "^1.2.1",
     "date-fns": "^2.30.0",
-    "framer-motion": "^10.12.18",
+    "framer-motion": "^10.16.4",
     "i18next": "^23.5.1",
     "i18next-browser-languagedetector": "^7.1.0",
     "i18next-resources-to-backend": "^1.1.4",
     "lodash": "^4.17.21",
-    "next": "13.4.19",
+    "next": "13.5.6",
     "react": "18.2.0",
     "react-datepicker": "^4.15.0",
     "react-dom": "18.2.0",
