@@ -1,4 +1,4 @@
-export type PageType = 'grouping' | 'meeting' | 'profile';
+export type PageType = 'grouping' | 'meeting' | 'community' | 'profile';
 
 export type StrictPropsWithChildren<P = unknown> = P & {
   children: React.ReactNode;
